apps/ai-assistant/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
accelerate==0.24.1
bitsandbytes==0.41.3
python-multipart==0.0.6
python-dotenv==1.0.0